Refactor time tracking in BasePlayer and related classes
- Updated `BasePlayer` constructor to not initialize qthread parent.
- Changed `elapsed` variable in `stopWatchedTiming` to `double` for precision.
- Modified return types of `getSessionTime` and `getTotalWatchedTime` to `double`, using `std::chrono::duration<double>`.
- Changed `totalWatchedTimeSeconds` in `BasePlayer` from `int` to `double`.
- Updated `incrementTimeWatchedTotal` and `incrementTimeSessionTotal` in `VideoWatcherQt` to accept `double` values.
- Ensured `resetDailyStats` correctly resets daily statistics.
- Enhanced `OnCopyData` in `MpcPlayer` for better command handling and signal emissions.
- Updated `sqliteDB` methods for incrementing time to accept `double` values.
- Adjusted header files for consistency with method signatures and member types.

+++ MediaManager/src/general/BasePlayer.h
@@ -31,7 +31,7 @@ public:
     QProcess* process;  
     std::shared_ptr<std::chrono::microseconds> sessionTimeStart = nullptr;  
     std::shared_ptr<std::chrono::microseconds> watchedTimeStart = nullptr;  
-    int totalWatchedTimeSeconds = 0;  
+    double totalWatchedTimeSeconds = 0.0;
     bool wasPlayingLastCheck = false;  
     BasePlayer(QString video_path, int video_id, int* CLASS_COUNT, QObject* parent = nullptr, MainApp* App = nullptr);  
     virtual ~BasePlayer();  
@@ -48,8 +48,8 @@ public:
     void stopSessionTiming();  
     void startWatchedTiming();  
     void stopWatchedTiming();  
-    int getSessionTime();  
-    int getTotalWatchedTime();  
+    double getSessionTime();  
+    double getTotalWatchedTime();  
     void updateWatchedTiming();
     void start();
 signals:
